TASK: ContentGraph bound to Workspace
The ContentGraph is now always bound to a
`WorkspaceName` and `ContentStreamId` combination.
The ProjectionState of the ContentGraph is now
the `ContentGraphFinder` which allows you to get
an instance of the `ContentGraph` bound to a
combination (usually by giving only the `WorkspaceName`).

The userland way is to use

`ContentRepository::getContentGraph(WorkspaceName)`