Move savepoint creation back on the subscription store
This reverts commit 82c69e70b1483814a69449499ab78fc629c52176.

We discussed that we dont want to ensure exactly once delivery for external projections. Adding transaction logic via traits for the projections increases logic there. And while with alot of effort we could bring exactly once delivery to work for most cases php could still decide to die in the small timeframe where we commit the one transaction and then the second. So there is no gurantee except when using a dedicated store: https://github.com/neos/neos-development-collection/pull/5377